Disallow mutable state in client decl. modules
