+++ compiler/ballerina-lang/src/main/resources/compiler.properties
@@ -1938,3 +1938,6 @@ error.unsupported.exposure.of.construct.from.module.generated.for.client.decl=\
 
 error.module.generated.for.client.decl.must.have.a.client.object.type=\
   a module generated for a client declaration must have an object type or class named ''client''
+
+error.module.generated.for.client.decl.cannot.have.mutable.state=\
+  a module generated for a client declaration cannot have mutable state
